Extract result-detail component from result component; refactor both components => constant naming, types, etc.
