+++ src/main/webapp/app/editor/instructions/editor-instructions-result-detail.ts
@@ -6,28 +6,29 @@ import {NgbActiveModal} from '@ng-bootstrap/ng-bootstrap';
 // Modal -> Result details view
 @Component({
     selector: 'jhi-editor-instructions-result-detail',
-    templateUrl: '../../entities/result/result-detail.html'
+    // This popup references the result detail html template, so make sure the constant names match
+    templateUrl: '../../entities/result/result-detail.component.html'
 })
 export class EditorInstructionsResultDetailComponent implements OnInit {
     @Input() result: Result;
-    @Input() tests;
-    loading: boolean;
-    filterTests: string;
-    details: Feedback[];
+    @Input() tests: string;
+    isLoading: boolean;
+    filterTests: string[];
+    feedbackList: Feedback[];
     buildLogs;
 
     constructor(public activeModal: NgbActiveModal,
                 private resultService: ResultService) {}
 
     ngOnInit(): void {
-        this.loading = true;
+        this.isLoading = true;
         this.filterTests = this.tests.split(',');
         this.resultService.details(this.result.id).subscribe(res => {
-            this.details = res.body.filter(
+            this.feedbackList = res.body.filter(
                 detail => this.filterTests.indexOf(detail.text) !== -1
             );
-            this.loading = false;
+            this.isLoading = false;
         });
-        this.loading = false;
+        this.isLoading = false;
     }
 }
