+++ src/main/webapp/app/entities/result/index.ts
@@ -1,4 +1,5 @@
 export * from './result.model';
 export * from './result.service';
 export * from './result.component';
+export * from './result-detail.component';
 export * from './result.module';
