+++ src/main/webapp/app/entities/result/result.module.ts
@@ -1,9 +1,7 @@
 import { CUSTOM_ELEMENTS_SCHEMA, NgModule } from '@angular/core';
-
 import { ArTEMiSSharedModule } from '../../shared';
-import { ResultService } from './';
+import { ResultComponent, ResultDetailComponent, ResultService } from './';
 import { ExerciseResultService } from './result.service';
-import { ResultComponent, ResultDetailComponent } from './result.component';
 import { MomentModule } from 'angular2-moment';
 
 @NgModule({
@@ -16,7 +14,8 @@ import { MomentModule } from 'angular2-moment';
         ResultDetailComponent
     ],
     exports: [
-        ResultComponent
+        ResultComponent,
+        ResultDetailComponent
     ],
     entryComponents: [
         ResultComponent,
